launching and other stuff

+++ Balatro Instance Manager/Form1.cs
@@ -1,12 +1,13 @@
-using System.IO;
+using System.Diagnostics;
 using Newtonsoft.Json;
-using Balatro_Instance_Manager.Properties;
+
 
 namespace Balatro_Instance_Manager
 {
     public partial class Form1 : Form
     {
         string balatroPathStr;
+        string balatroSteamDirStr;
         public Form1()
         {
             InitializeComponent();
@@ -20,7 +21,14 @@ namespace Balatro_Instance_Manager
                 balatroPathStr = Path.Combine(Environment.GetFolderPath(Environment.SpecialFolder.ApplicationData), "Balatro");
             }
 
+            balatroSteamDirStr = Properties.Settings.Default.BalatroSteamPath;
+            if (string.IsNullOrEmpty(balatroPathStr))
+            {
+                balatroSteamDirStr = "C:\\Program Files (x86)\\Steam\\steamapps\\common\\Balatro";
+            }
+
             balatroPath.Text = balatroPathStr;
+            balatroSteamPath.Text = balatroSteamDirStr;
             loadProfiles();
         }
 
@@ -33,7 +41,7 @@ namespace Balatro_Instance_Manager
 
             if (!IsDirectoryAccessible(balatroPath.Text))
             {
-                MessageBox.Show("Balatro path is not accessible. Please check the path and try again.", "Information", MessageBoxButtons.OK);
+                MessageBox.Show("Balatro data path is not accessible. Please check the path and try again.", "Information", MessageBoxButtons.OK);
                 return;
             }
 
@@ -60,7 +68,7 @@ namespace Balatro_Instance_Manager
             string modProfilesPath = Path.Combine(balatroPathStr, "ModProfiles");
             if (!IsDirectoryAccessible(modProfilesPath))
             {
-                MessageBox.Show("Balatro path is not accessible, make sure to check the path before launching.", "Information", MessageBoxButtons.OK);
+                MessageBox.Show("Balatro data path is not accessible, make sure to check the path before launching.", "Information", MessageBoxButtons.OK);
                 return;
             }
 
@@ -72,7 +80,8 @@ namespace Balatro_Instance_Manager
             foreach (var subdirectory in subdirectories)
             {
                 string profilePath = Path.Combine(subdirectory, "profile.json");
-                File.WriteAllText(profilePath, "{\"Enabled\": false}");
+                if (!File.Exists(profilePath))
+                    File.WriteAllText(profilePath, "{\"Enabled\": false}");
 
                 try
                 {
@@ -82,6 +91,7 @@ namespace Balatro_Instance_Manager
                     CheckBox checkBox = new CheckBox();
                     checkBox.Text = Path.GetFileName(subdirectory);
                     checkBox.Checked = profile.Enabled;
+                    checkBox.CheckedChanged += onCheckClicked;
 
                     instances.Controls.Add(checkBox);
                 }
@@ -114,12 +124,311 @@ namespace Balatro_Instance_Manager
                 return false;
             }
         }
+
+        private void onCheckClicked(object? sender, EventArgs e)
+        {
+            if (sender is CheckBox checkBox)
+            {
+                string profilePath = Path.Combine(balatroPathStr, "ModProfiles", checkBox.Text, "profile.json");
+                try
+                {
+                    string jsonContent = File.ReadAllText(profilePath);
+                    var profile = JsonConvert.DeserializeObject<Profile>(jsonContent);
+                    if (profile != null)
+                    {
+                        profile.Enabled = checkBox.Checked;
+                        File.WriteAllText(profilePath, JsonConvert.SerializeObject(profile));
+                    }
+                }
+                catch (Exception ex)
+                {
+                    Console.WriteLine($"Error reading JSON in {profilePath}: {ex.Message}");
+                }
+            }
+        }
+
+        private void balatroSteamPath_KeyDown(object sender, KeyEventArgs e)
+        {
+            if (e.KeyCode == Keys.Escape || e.KeyCode == Keys.Enter)
+            {
+                this.ActiveControl = null;
+                e.Handled = true;
+                e.SuppressKeyPress = true;
+            }
+        }
+
+        private void balatroSteamPath_Unfocused(object sender, EventArgs e)
+        {
+            if (balatroSteamPath.Text == "")
+            {
+                balatroSteamPath.Text = "C:\\Program Files (x86)\\Steam\\steamapps\\common\\Balatro";
+            }
+
+            if (!IsDirectoryAccessible(balatroSteamPath.Text))
+            {
+                MessageBox.Show("Balatro steam path is not accessible. Please check the path and try again.", "Information", MessageBoxButtons.OK);
+                return;
+            }
+
+            balatroSteamDirStr = balatroSteamPath.Text;
+
+            Properties.Settings.Default.BalatroSteamPath = balatroSteamDirStr;
+            Properties.Settings.Default.Save();
+
+            loadProfiles();
+        }
+
+        private void launch_Click(object sender, EventArgs e)
+        {
+            if (!IsDirectoryAccessible(balatroSteamDirStr))
+            {
+                MessageBox.Show("Balatro steam path is not accessible. Please check the path and try again.", "Information", MessageBoxButtons.OK);
+                return;
+            }
+
+            string modProfilesPath = Path.Combine(balatroPathStr, "ModProfiles");
+            if (!IsDirectoryAccessible(modProfilesPath))
+            {
+                MessageBox.Show("Balatro data path is not accessible, make sure to check the path before launching.", "Information", MessageBoxButtons.OK);
+                return;
+            }
+
+            List<string> mods = GetProfileMods();
+            if (mods == null) return;
+
+            HashSet<string> loadedModIds = new HashSet<string>();
+            List<string> validMods = new List<string>();
+
+            foreach (var mod in mods)
+            {
+                var jsonFiles = Directory.GetFiles(mod, "*.json");
+
+                bool isValidMod = true;
+
+                foreach (var jsonFile in jsonFiles)
+                {
+                    try
+                    {
+                        string jsonContent = File.ReadAllText(jsonFile);
+                        var modData = JsonConvert.DeserializeObject<ModData>(jsonContent);
+
+                        if (modData != null)
+                        {
+                            if (!string.IsNullOrEmpty(modData.id))
+                            {
+                                if (loadedModIds.Contains(modData.id))
+                                {
+                                    isValidMod = false;
+                                    break;
+                                }
+                            }
+                            else if (!string.IsNullOrEmpty(modData.name))
+                            {
+                                if (loadedModIds.Contains(modData.name))
+                                {
+                                    isValidMod = false;
+                                    break;
+                                }
+                            }
+                        }
+                    }
+                    catch (Exception ex)
+                    {
+                        Console.WriteLine($"Error reading JSON in {jsonFile}: {ex.Message}");
+                        isValidMod = false;
+                        break;
+                    }
+                }
+
+                if (isValidMod)
+                {
+                    validMods.Add(mod);
+                    foreach (var jsonFile in jsonFiles)
+                    {
+                        try
+                        {
+                            string jsonContent = File.ReadAllText(jsonFile);
+                            var modData = JsonConvert.DeserializeObject<ModData>(jsonContent);
+
+                            if (modData != null)
+                            {
+                                if (!string.IsNullOrEmpty(modData.id))
+                                {
+                                    loadedModIds.Add(modData.id);
+                                }
+                                else if (!string.IsNullOrEmpty(modData.name))
+                                {
+                                    loadedModIds.Add(modData.name);
+                                }
+                            }
+                        }
+                        catch (Exception ex)
+                        {
+                            Console.WriteLine($"Error reading JSON in {jsonFile}: {ex.Message}");
+                        }
+                    }
+                }
+            }
+
+
+            string modsDirectory = Path.Combine(balatroPathStr, "Mods");
+            if (!Directory.Exists(modsDirectory))
+            {
+                MessageBox.Show("Mods directory does not exist. Make sure you have lovely installed and have opened the game at least once since then and try again.", "Error", MessageBoxButtons.OK);
+                return;
+            }
+
+            if(deselected.Checked) ClearProfileMods();
+
+            foreach (var mod in validMods)
+            {
+                string output = Path.Combine(modsDirectory, Path.GetFileName(mod));
+                CopyDirectory(mod, output);
+                File.WriteAllText(Path.Combine(output, ".bim_profile"), "");
+                if (lovelyIgnore.Checked)
+                {
+                    string lovelyIgnorePath = Path.Combine(output, ".lovelyignore");
+                    if (File.Exists(lovelyIgnorePath))
+                    {
+                        try
+                        {
+                            File.Delete(lovelyIgnorePath);
+                        }
+                        catch (Exception)
+                        {
+                            continue;
+                        }
+                    }
+                }
+            }
+
+
+            string exePath = Path.Combine(balatroSteamDirStr, "Balatro.exe");
+            if (File.Exists(exePath))
+            {
+                Process.Start(exePath);
+            }
+            else
+            {
+                MessageBox.Show("Executable not found. Please check the path and try again.", "Error", MessageBoxButtons.OK);
+            }
+        }
+        private List<string> GetProfileMods()
+        {
+            string modProfilesPath = Path.Combine(balatroPathStr, "ModProfiles");
+            if (!IsDirectoryAccessible(modProfilesPath)) return null;
+
+            Directory.CreateDirectory(modProfilesPath);
+
+            var subdirectories = Directory.GetDirectories(modProfilesPath);
+
+            List<string> mods = new List<string>();
+
+            foreach (var subdirectory in subdirectories)
+            {
+                string profilePath = Path.Combine(subdirectory, "profile.json");
+                if (!File.Exists(profilePath))
+                    File.WriteAllText(profilePath, "{\"Enabled\": false}");
+
+                try
+                {
+                    string jsonContent = File.ReadAllText(profilePath);
+                    var profile = JsonConvert.DeserializeObject<Profile>(jsonContent);
+
+                    if (!profile.Enabled) continue;
+                }
+                catch (Exception ex)
+                {
+                    Console.WriteLine($"Error reading JSON in {subdirectory}: {ex.Message}");
+                    continue;
+                }
+
+                var modsInProfile = Directory.GetDirectories(subdirectory);
+                foreach (var mod in modsInProfile)
+                {
+                    mods.Add(mod);
+                }
+            }
+
+            return mods;
+        }
+
+        private void CopyDirectory(string sourceDir, string destDir)
+        {
+            Directory.CreateDirectory(destDir);
+
+            foreach (var file in Directory.GetFiles(sourceDir))
+            {
+                string destFile = Path.Combine(destDir, Path.GetFileName(file));
+                File.Copy(file, destFile, true);
+            }
+
+            foreach (var subdir in Directory.GetDirectories(sourceDir))
+            {
+                string destSubdir = Path.Combine(destDir, Path.GetFileName(subdir));
+                CopyDirectory(subdir, destSubdir);
+            }
+        }
+
+        private void ClearProfileMods()
+        {
+            string modsDirectory = Path.Combine(balatroPathStr, "Mods");
+            if (!Directory.Exists(modsDirectory)) return;
+
+            var modFolders = Directory.GetDirectories(modsDirectory);
+            foreach (var modFolder in modFolders)
+            {
+                string bimProfilePath = Path.Combine(modFolder, ".bim_profile");
+                if (File.Exists(bimProfilePath))
+                {
+                    try
+                    {
+                        Directory.Delete(modFolder, true);
+                    }
+                    catch (Exception)
+                    {
+                        continue;
+                    }
+                }
+            }
+        }
+
+        private void clear_Click(object sender, EventArgs e)
+        {
+            string modProfilesPath = Path.Combine(balatroPathStr, "ModProfiles");
+            if (!IsDirectoryAccessible(modProfilesPath))
+            {
+                MessageBox.Show("Balatro data path is not accessible, make sure to check the path before launching.", "Information", MessageBoxButtons.OK);
+                return;
+            }
+
+            if (!IsDirectoryAccessible(balatroSteamDirStr))
+            {
+                MessageBox.Show("Balatro steam path is not accessible. Please check the path and try again.", "Information", MessageBoxButtons.OK);
+                return;
+            }
+
+            string modsDirectory = Path.Combine(balatroPathStr, "Mods");
+            if (!Directory.Exists(modsDirectory))
+            {
+                MessageBox.Show("Mods directory does not exist. Make sure you have lovely installed and have opened the game at least once since then and try again.", "Error", MessageBoxButtons.OK);
+                return;
+            }
+
+            ClearProfileMods();
+            MessageBox.Show("Mods cleared successfully.", "Information", MessageBoxButtons.OK);
+        }
     }
 
 }
 
 public class Profile
 {
-    public string Name { get; set; }
     public bool Enabled { get; set; }
+}
+
+public class ModData
+{
+    public string id { get; set; }
+    public string name { get; set; }
 }
